Fix: Broken CLI (vpnctl) didn't work (#16)
This PR restores the basics of the `vpnctl` CLI functionality. The `show` and `list` commands work for all levels. The `add`, `edit`, `delete` and `commit` commands work on the upper level commands as was the case previously.

A `summary` command has been added to most levels to show the live status of connections.

Values are now have autocompletion as well, making it easier to determine what to pass.

Documentation in the readme's will be completed in another PR