Feature: add SSH tunnel capability (#18)
This PR adds the ability to create SSH tunnels in a network instance. SSH tunnels can only be created within an instance and as such require another connection in the instance to work. This is handy in cases where an IPsec or physical connection is available, but only a route to a single jump host is allowed. SSH can then be used to tunnel traffic further into the network.